feat #26: TokenService 인터페이스 구현 후 의존 변경
